requirements.txt
astroid==2.1.0
certifi==2018.11.29
Click==7.0
colorama==0.4.1
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.1.0
mccabe==0.6.1
pylint==2.2.2
pyodbc==4.0.25
python-dotenv==0.10.1
six==1.12.0
SQLAlchemy==1.2.16
Werkzeug==0.14.1
wincertstore==0.2
wrapt==1.11.1
